修改

+++ TiHMS/CommonModule/src/main/java/tiquartet/CommonModule/vo/OrderStrategyVO.java
@@ -1,8 +1,12 @@
 package tiquartet.CommonModule.vo;
 
 public class OrderStrategyVO {
+	 //订单编号
      public long orderID;
+     //策略介绍
      public String strategyIntroduce;
-     public int orderPrice;
-     public long strategyID;
+     //订单价格
+     public double orderPrice;
+     //策略编号
+     public int strategyID;
 }
